Removed custom CMake file. Moved demos out of this project. Moved non-arch files into base source/header folders. Added KUBOS_CORE_XXX ifdef into arch source files.
