+++ kubos-core/k_alloc.h
