+++ source/arch/posix/k_uart_linux.c
@@ -18,6 +18,8 @@ License along with this library; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
+#ifdef KUBOS_CORE_POSIX
+
 #include "kubos-core/arch/k_uart.h"
 
 #include <stdio.h>
@@ -205,3 +207,5 @@ static void *serial_rx_thread(void *vptr_args) {
 	}
 	return NULL;
 }
+
+#endif
